pyproject.toml
[tool.poetry]
name = "zurkon"
version = "0.1.0"
description = ""
authors = ["cahillross"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
discord-py = "^2.3.2"
pynacl = "^1.5.0"
yt-dlp = "^2023.12.30"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
black = "^23.12.1"
isort = "^5.13.2"
mypy = "^1.8.0"
bandit = "^1.7.6"
pre-commit = "^3.6.0"
pytest = "^7.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.setuptools.packages.find]
where = ["zurkon"]